refactor(config): move strategy classes to config model directory
