+++ src/Interface/Config.php
@@ -2,18 +2,18 @@
 
 namespace Brainshaker95\PhpToTsBundle\Interface;
 
-use Brainshaker95\PhpToTsBundle\FileNameStrategy\KebabCase;
-use Brainshaker95\PhpToTsBundle\Model\Config\FileType as ConfigFileType;
-use Brainshaker95\PhpToTsBundle\Model\Config\Indent as ConfigIndent;
-use Brainshaker95\PhpToTsBundle\SortStrategy\AlphabeticalAsc;
-use Brainshaker95\PhpToTsBundle\SortStrategy\ConstructorFirst;
-use Brainshaker95\PhpToTsBundle\SortStrategy\ReadonlyFirst;
+use Brainshaker95\PhpToTsBundle\Model\Config\FileNameStrategy\KebabCase;
+use Brainshaker95\PhpToTsBundle\Model\Config\FileType;
+use Brainshaker95\PhpToTsBundle\Model\Config\Indent;
+use Brainshaker95\PhpToTsBundle\Model\Config\SortStrategy\AlphabeticalAsc;
+use Brainshaker95\PhpToTsBundle\Model\Config\SortStrategy\ConstructorFirst;
+use Brainshaker95\PhpToTsBundle\Model\Config\SortStrategy\ReadonlyFirst;
 
 interface Config
 {
-    public const DEFAULT_FILE_TYPE          = ConfigFileType::TYPE_MODULE;
+    public const DEFAULT_FILE_TYPE          = FileType::TYPE_MODULE;
     public const DEFAULT_INDENT_COUNT       = 2;
-    public const DEFAULT_INDENT_STYLE       = ConfigIndent::STYLE_SPACE;
+    public const DEFAULT_INDENT_STYLE       = Indent::STYLE_SPACE;
     public const DEFAULT_INPUT_DIR          = 'src/Model/TypeScriptables';
     public const DEFAULT_OUTPUT_DIR         = 'resources/ts/types/generated';
     public const DEFAULT_FILE_NAME_STRATEGY = KebabCase::class;
@@ -29,11 +29,11 @@ interface Config
     public function getOutputDir(): ?string;
 
     /**
-     * @phpstan-return ConfigFileType::TYPE_*
+     * @phpstan-return FileType::TYPE_*
      */
     public function getFileType(): ?string;
 
-    public function getIndent(): ?ConfigIndent;
+    public function getIndent(): ?Indent;
 
     /**
      * @return class-string<SortStrategy>[]
