+++ src/Model/Config/FileNameStrategy/KebabCase.php
@@ -1,14 +1,14 @@
 <?php
 
-namespace Brainshaker95\PhpToTsBundle\FileNameStrategy;
+namespace Brainshaker95\PhpToTsBundle\Model\Config\FileNameStrategy;
 
 use Brainshaker95\PhpToTsBundle\Interface\FileNameStrategy;
 use Brainshaker95\PhpToTsBundle\Tool\Str;
 
-class PascalCase implements FileNameStrategy
+class KebabCase implements FileNameStrategy
 {
     public function getName(string $name): string
     {
-        return Str::toPascal($name);
+        return Str::toKebab($name);
     }
 }
