+++ src/Model/Config/FileNameStrategy/LowerCase.php
@@ -1,14 +1,14 @@
 <?php
 
-namespace Brainshaker95\PhpToTsBundle\FileNameStrategy;
+namespace Brainshaker95\PhpToTsBundle\Model\Config\FileNameStrategy;
 
 use Brainshaker95\PhpToTsBundle\Interface\FileNameStrategy;
 use Brainshaker95\PhpToTsBundle\Tool\Str;
 
-class KebabCase implements FileNameStrategy
+class LowerCase implements FileNameStrategy
 {
     public function getName(string $name): string
     {
-        return Str::toKebab($name);
+        return Str::toLower($name);
     }
 }
