+++ src/Model/Config/FileNameStrategy/PascalCase.php
@@ -1,14 +1,14 @@
 <?php
 
-namespace Brainshaker95\PhpToTsBundle\FileNameStrategy;
+namespace Brainshaker95\PhpToTsBundle\Model\Config\FileNameStrategy;
 
 use Brainshaker95\PhpToTsBundle\Interface\FileNameStrategy;
 use Brainshaker95\PhpToTsBundle\Tool\Str;
 
-class LowerCase implements FileNameStrategy
+class PascalCase implements FileNameStrategy
 {
     public function getName(string $name): string
     {
-        return Str::toLower($name);
+        return Str::toPascal($name);
     }
 }
