+++ src/Model/Config/FileNameStrategy/SnakeCase.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Brainshaker95\PhpToTsBundle\FileNameStrategy;
+namespace Brainshaker95\PhpToTsBundle\Model\Config\FileNameStrategy;
 
 use Brainshaker95\PhpToTsBundle\Interface\FileNameStrategy;
 use Brainshaker95\PhpToTsBundle\Tool\Str;
