+++ src/Model/Config/SortStrategy/AlphabeticalDesc.php
@@ -1,14 +1,14 @@
 <?php
 
-namespace Brainshaker95\PhpToTsBundle\SortStrategy;
+namespace Brainshaker95\PhpToTsBundle\Model\Config\SortStrategy;
 
 use Brainshaker95\PhpToTsBundle\Interface\SortStrategy;
 use Brainshaker95\PhpToTsBundle\Model\TsProperty;
 
-class AlphabeticalAsc implements SortStrategy
+class AlphabeticalDesc implements SortStrategy
 {
     public function sort(TsProperty $property1, TsProperty $property2): int
     {
-        return $property1->name <=> $property2->name;
+        return $property2->name <=> $property1->name;
     }
 }
