+++ src/Model/Config/SortStrategy/AsDefined.php
@@ -1,14 +1,14 @@
 <?php
 
-namespace Brainshaker95\PhpToTsBundle\SortStrategy;
+namespace Brainshaker95\PhpToTsBundle\Model\Config\SortStrategy;
 
 use Brainshaker95\PhpToTsBundle\Interface\SortStrategy;
 use Brainshaker95\PhpToTsBundle\Model\TsProperty;
 
-class ReadonlyLast implements SortStrategy
+class AsDefined implements SortStrategy
 {
     public function sort(TsProperty $property1, TsProperty $property2): int
     {
-        return $property1->isReadonly <=> $property2->isReadonly;
+        return 0;
     }
 }
