+++ src/Model/Config/SortStrategy/ConstructorFirst.php
@@ -1,14 +1,14 @@
 <?php
 
-namespace Brainshaker95\PhpToTsBundle\SortStrategy;
+namespace Brainshaker95\PhpToTsBundle\Model\Config\SortStrategy;
 
 use Brainshaker95\PhpToTsBundle\Interface\SortStrategy;
 use Brainshaker95\PhpToTsBundle\Model\TsProperty;
 
-class AlphabeticalDesc implements SortStrategy
+class ConstructorFirst implements SortStrategy
 {
     public function sort(TsProperty $property1, TsProperty $property2): int
     {
-        return $property2->name <=> $property1->name;
+        return $property2->isConstructorProperty <=> $property1->isConstructorProperty;
     }
 }
