+++ src/Model/Config/SortStrategy/ReadonlyFirst.php
@@ -1,14 +1,14 @@
 <?php
 
-namespace Brainshaker95\PhpToTsBundle\SortStrategy;
+namespace Brainshaker95\PhpToTsBundle\Model\Config\SortStrategy;
 
 use Brainshaker95\PhpToTsBundle\Interface\SortStrategy;
 use Brainshaker95\PhpToTsBundle\Model\TsProperty;
 
-class ConstructorLast implements SortStrategy
+class ReadonlyFirst implements SortStrategy
 {
     public function sort(TsProperty $property1, TsProperty $property2): int
     {
-        return $property1->isConstructorProperty <=> $property2->isConstructorProperty;
+        return $property2->isReadonly <=> $property1->isReadonly;
     }
 }
